Removed attempt to include Template.ini
Honestly... manifests are sooo poorly documented.
And, more importantly ... i didn't HAVE to include/define this file as something to include.
(in fact, i removed it because trying to PUBLISH the module with this directive BUSTED the module's publication - but removing the directive ... a) made the build work and b) (of course) the file 'goes into' the build ... so, screw it...)