PRO224: MAJOR REFACTOR
New 'verb' for assignment of ProvisoRoot: Assign.
With => Target (and only targets configs/machines).
Assign (which used to TRY and do ProvisoRoot and DomainCreds and Security/Secrets) now JUST does domain creds and sensitive info and ... => Map (with alias of Add).

Further, With (now Target) is no longer 'bound' to the pipeline/required as a call before EVERY Configure|Validate call. it's a 'floater' or part of setup. HAS to be called initially, but ... way less tedious.